+++ CHANGELOG.md
@@ -8,10 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Known Issues
 - Reinstalling the mod may sometimes cause modded language data to be deleted.
 
-# 2.2.1 (2025-01-??)
-This update provides some updated language strings as well as a bugfix.
+# 2.3.0 (2025-07-??)
+This update updates the game for provides some updated language strings as well as a bugfix.
 
 ## Added
+- Added cause of death for the new 
 - Added additional death messages for English (via Thundermaker300).
 # Changed
 - Updated Italian translations (via MakinDay)